day2_streamlit_step0/streamlit/requirements.txt
altair
pandas
streamlit
pymongo
plotly-express